edit : api 호출 로직 변경

+++ infrastructure/src/main/kotlin/com/searchblog/infrastructure/api/dto/NaverBlogResponse.kt
@@ -5,9 +5,9 @@ import com.fasterxml.jackson.annotation.JsonProperty
 class NaverBlogResponse(
   val lastBuildDate: String,
 
-  val total: Long,
+  val total: Int,
 
-  val start: Long,
+  val start: Int,
 
   val display: Long,
